🐛fix: NavBar 설정수정
